# controller/src/main/resources/public/api.yaml
openapi: 3.0.1
info:
  title: Car Hire Rental API
  version: 0.1.0
  contact:
    name: Adam Alexandru
  description: Car Hire Rental API
servers:
  - url: http://localhost:{port}
    description: local environment
    variables:
      port:
        default: '9090'
tags:
  - name: rental
    description: Rental service
security:
  - oAuth:
      - write
      - read
paths:
  /rent:
    post:
      tags:
        - rental
      summary: Create new reservation
      operationId: createReservation
      requestBody:
        description: "Create reservation body"
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NewReservationDTO"
      responses:
        '201':
          $ref: '#/components/responses/ResourceCreated'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/ServerError'
    get:
      tags:
        - rental
      summary: Get all reservations
      description: Return a list with all the reservations
      operationId: getAllReservations
      parameters:
        - in: query
          name: status
          required: false
          schema:
            $ref: "#/components/schemas/StatusEnum"
          description: status of the reservation
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: "array"
                items:
                  $ref: "#/components/schemas/ReservationDTO"
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
  /rent/{id}:
    get:
      tags:
        - rental
      summary: Get reservation by ID
      description: Return the reservation found by ID
      operationId: getReservationByID
      parameters:
        - in: "path"
          name: "id"
          required: true
          schema:
            type: "string"
          description: Reservation ID
      responses:
        '200':
          description: Return user's details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReservationDTO'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
    patch:
      tags:
        - rental
      summary: Update reservation details
      description: Update reservation details with new information
      operationId: updateReservationDetails
      parameters:
        - in: "path"
          name: "id"
          required: true
          schema:
            type: "string"
          description: Reservation ID
      requestBody:
        description: "Update reservation body"
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PatchBody"
      responses:
        '200':
          $ref: '#/components/responses/Success'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/ServerError'
    delete:
      tags:
        - rental
      summary: Delete reservation
      description: Delete reservation from database
      operationId: deleteReservation
      parameters:
        - in: "path"
          name: "id"
          required: true
          schema:
            type: "string"
          description: Reservation ID
      responses:
        '200':
          $ref: '#/components/responses/Success'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'


components:
  securitySchemes:
    oAuth:
      type: oauth2
      description: This API uses OAuth2
      flows:
        password:
          tokenUrl: /oauth/token
          scopes:
            read: Get information from the API
            write: Use the API to create or update data
  responses:
    ResourceCreated:
      description: success
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ResourceCreatedDTO'
    Success:
      description: success
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ResponseDTO'
    Unauthorized:
      description: unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorDTO'
    NotFound:
      description: not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorDTO'
    BadRequest:
      description: bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorDTO'
    ServerError:
      description: internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorDTO'
  schemas:
    OpEnum:
      type: string
      description: Operation available for object
      enum:
        - ADD
        - REMOVE
        - REPLACE
    PatchBody:
      type: array
      items:
        $ref: '#/components/schemas/PatchDocument'
    PatchDocument:
      type: object
      description: HTTP JSONPatch document from RFC 6902
      properties:
        op:
          $ref: "#/components/schemas/OpEnum"
        path:
          type: string
          description: JSON Pointer
          example: "a/b/c"
        value:
          type: string
          description: JSON pointer value
          example: "newValue"
      required:
        - op
        - path
        - value

    StatusEnum:
      type: string
      enum:
        - Active
        - Deleted

    CarClassEnum:
      type: string
      enum:
        - Coupe
        - Hatchback
        - Limousine
        - Sedan
        - Minivan
        - Suv

    NewUserDTO:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
          format: "password"
        firstname:
          type: string
        lastname:
          type: string
        countryOfResidence:
          type: string
          example: "ID of country"
        driverLicense:
          type: string
    UserDTO:
      type: object
      properties:
        id:
          type: string
        username:
          type: string
        firstname:
          type: string
        lastname:
          type: string
        countryOfResidence:
          type: string
        driverLicense:
          type: string
        status:
          $ref: "#/components/schemas/StatusEnum"
    NewCarDTO:
      type: object
      properties:
        make:
          type: string
          example: "ID of make"
        model:
          type: string
        year:
          type: integer
        mileage:
          type: integer
        fuel:
          type: number
          format: float
        carClass:
          $ref: "#/components/schemas/CarClassEnum"
    CarDTO:
      type: object
      properties:
        id:
          type: string
        make:
          type: string
        model:
          type: string
        year:
          type: integer
        mileage:
          type: integer
        fuel:
          type: number
          format: float
        carClass:
          $ref: "#/components/schemas/CarClassEnum"
        status:
          $ref: "#/components/schemas/StatusEnum"
    NewReservationDTO:
      type: object
      properties:
        userID:
          type: string
          example: "User ID"
        carID:
          type: string
          example: "Car ID"
        startDateTime:
          type: string
          format: "date-time"
        endDateTime:
          type: string
          format: "date-time"
    ReservationDTO:
      type: object
      properties:
        id:
          type: string
          example: "Reservation ID"
        userID:
          type: string
          example: "User ID"
        carID:
          type: string
          example: "Car ID"
        startDateTime:
          type: string
          format: "date-time"
        endDateTime:
          type: string
          format: "date-time"
        status:
          $ref: "#/components/schemas/StatusEnum"
    ResourceCreatedDTO:
      type: object
      properties:
        id:
          type: string
          description: ID of the created resource
    ErrorDTO:
      type: object
      properties:
        code:
          type: integer
          example: 400
        message:
          type: string
          example: error message
    ResponseDTO:
      type: object
      properties:
        message:
          type: string
          example: Sample message
          description: Custom message